Split test classes for default and custom demo guards
Add trait with shared tests for both test classes
Add custom guard test class
